Created earthquake prediction visualization by combining previous plots.
Combined multiple Point plots to create one interactive plot. Points in
geoviews can not have legends, so I added a selector to highlight the
past vs future points. Instead of a selector we can add markdown to the
right of the figure explaining the color of past vs future points.